test/stepper_motor_test/requirements.txt
pyserial>=3.5
matplotlib>=3.8.0
